requirements.txt
aenum==2.2.3
apturl==0.5.2
astroid==2.4.1
attrs==19.3.0
beautifulsoup4==4.9.1
blinker==1.4
Brlapi==0.7.0
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
dialogflow==1.0.0
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
fbchat==1.9.7
Flask==1.1.2
google-api-core==1.20.0
google-auth==1.17.2
googleapis-common-protos==1.52.0
grpcio==1.29.0
httplib2==0.14.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kazam==1.4.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
openshot-qt==2.5.1.dev2
paho-mqtt==1.5.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.5.2
pymacaroons==0.13.0
PyNaCl==1.3.0
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
reportlab==3.5.34
requests==2.23.0
requests-unixsocket==0.2.0
rsa==4.6
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==2.0.1
systemd-python==234
toml==0.10.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
wadllib==1.3.3
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
